Finished propagating new project structure changes through analysis code (see previous commit). Code compiles and analysis proceeds at least through histos() with no obvious bugs. Cleaned up Config.h/.cpp and abstracted the 4M/6M detector arms so that an arbitrary number of detectors could be used through at least subtractBackground().
Note that DETECTOR_NAMES is defined to include *all* detectors that provided
data in the experiment, not just the 4M/6M arms from which cross sections are
calculated. May want to define a new config file for these, but for now these
names are part of the experimentalConstants.h file.

Need to finish development in calculateReference.cpp to allow for consistent
treatment of reference runs (e.g., polyethylene/graphite/blank).